+++ src/index.js
@@ -1,10 +1,12 @@
 import TmHelloWorld from "./components/tmHelloWorld"
 import TmField from "./components/tmField"
+import TmPageHeader from "./components/tmPageHeader"
 import TmNotifications from "./components/tmNotifications"
 
 const TendermintComponents = {
   TmHelloWorld,
   TmField,
+  TmPageHeader,
   TmNotifications
 }
 
@@ -21,5 +23,7 @@ export {
 } from "./components/tmHelloWorld/tmHelloWorld.vue"
 export { default as TmField } from "./components/tmField/tmField.vue"
 export {
+  default as TmPageHeader
+} from "./components/tmPageHeader/tmPageHeader.vue"
   default as TmNotifications
 } from "./components/tmNotifications/tmNotifications.vue"