+++ tests/unit/tmPageHeader.spec.js
@@ -1,0 +1,21 @@
+import { shallowMount } from "@vue/test-utils"
+import htmlBeautify from "html-beautify"
+import { TmPageHeader } from "@/index.js"
+
+describe("TmPageHeader.vue", () => {
+  let wrapper
+  beforeEach(async () => {
+    const title = "Title"
+    const subtitle = "Subtitle"
+    wrapper = shallowMount(TmPageHeader, {
+      propsData: { title, subtitle },
+      slots: {
+        default: "<div>slot</div>"
+      }
+    })
+  })
+
+  it("has the expected html structure", () => {
+    expect(htmlBeautify(wrapper.html())).toMatchSnapshot()
+  })
+})